Change buflen and wirelen arguments of bpfjit_execute_code to size_t.
The change makes a difference on 64bit arches. For instance,
this is how the function looks like on amd64 before and after
the change:

0000000000001ac5 <bpfjit_execute_code>:
    1ac5:       48 83 ec 08             sub    $0x8,%rsp
    1ac9:       89 d2                   mov    %edx,%edx
    1acb:       89 f6                   mov    %esi,%esi
    1acd:       ff d1                   callq  *%rcx
    1acf:       48 83 c4 08             add    $0x8,%rsp
    1ad3:       c3                      retq

0000000000001ac5 <bpfjit_execute_code>:
    1ac5:       48 83 ec 08             sub    $0x8,%rsp
    1ac9:       ff d1                   callq  *%rcx
    1acb:       48 83 c4 08             add    $0x8,%rsp
    1acf:       c3                      retq

+++ src/bpfjit.c
@@ -957,8 +957,8 @@ fail:
 }
 
 unsigned int
-bpfjit_execute_code(const uint8_t *p, unsigned int wirelen,
-    unsigned int buflen, const void *code)
+bpfjit_execute_code(const uint8_t *p, size_t wirelen,
+    size_t buflen, const void *code)
 {
 	union
 	{
